saved up a whole bunch

+++ proj1/DataLine.cpp
@@ -5,7 +5,7 @@
 #include "DataLine.h"
 
 
-DataLine::DataLine(vector<float> feature_vector_in, string classification_in) {
+DataLine::DataLine(vector<int> feature_vector_in, string classification_in) {
     feature_vector = feature_vector_in;
     classification = classification_in;
 }
@@ -14,6 +14,6 @@ string DataLine::get_classification() {
     return classification;
 }
 
-vector<float> DataLine::get_feature_vector() {
+vector<int> DataLine::get_feature_vector() {
     return feature_vector;
 }
